+++ lib/forest_liana/mixpanel_event.rb
@@ -1,0 +1,4 @@
+module ForestLiana
+  class MixpanelEvent < OpenStruct
+  end
+end